voice-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch>=2.0.0
numpy>=1.24.0
soundfile>=0.12.0
python-dotenv==1.0.0
openai==1.12.0
gunicorn==21.2.0



